fix some bugs

+++ Tensorflow/Model/CNN.py
@@ -168,7 +168,7 @@ class CNN():
             flat_output_shape = [output_shape[0], output_shape[1] * output_shape[2] * output_shape[3]]
 
             pool_ = tf.reshape(input, [flat_input_size])
-            batch_range = tf.reshape(tf.range(tf.cast(output_shape[0], tf.float32), dtype=ind.dtype), 
+            batch_range = tf.reshape(tf.range(tf.cast(output_shape[0], tf.int64), dtype=ind.dtype), 
                                                 shape=[input_shape[0], 1, 1, 1])
             b = tf.ones_like(ind) * batch_range
             b1 = tf.reshape(b, [flat_input_size, 1])
